s3-worker/requirements.txt
fastapi
python-multipart
uvicorn
boto3
botocore
python-dotenv